+++ app/domain/token_factory.rb
@@ -9,22 +9,28 @@ class TokenFactory < Dry::Struct
 
   attribute :slosilo, ::Types::Any.default{ Slosilo }
 
+  MAXIMUM_AUTHENTICATION_TOKEN_EXPIRATION = 5.hours
+
   def signing_key(account)
     slosilo["authn:#{account}".to_sym] || raise(NoSigningKey, account)
   end
 
-  def signed_host_key(account:, username:)
-    exp_t = Time.now + Rails.application.config.conjur_config.host_authorization_token_ttl.to_i
-    signing_key(account).issue_jwt(sub: username, exp: exp_t)
+  def signed_token(account:,
+                   username:,
+                   host_ttl: Rails.application.config.conjur_config.host_authorization_token_ttl,
+                   user_ttl: Rails.application.config.conjur_config.user_authorization_token_ttl)
+    signing_key(account).issue_jwt(
+      sub: username,
+      exp: Time.now + offset(
+        ttl: username.starts_with?('host/') ? host_ttl : user_ttl
+      )
+    )
   end
 
-  def signed_token(account:, username:)
-    if username.starts_with?('host/')
-      signed_host_key(account: account, username: username)
-    else
-      exp_t = Time.now + Rails.application.config.conjur_config.user_authorization_token_ttl.to_i
-      signing_key(account).issue_jwt(sub: username, exp: exp_t)
-    end
+  def offset(ttl:)
+    return ttl if ttl < MAXIMUM_AUTHENTICATION_TOKEN_EXPIRATION
+
+    MAXIMUM_AUTHENTICATION_TOKEN_EXPIRATION
   end
 
 end